refactor: reorganize badge components and update imports for consistency
