+++ clients/assessment_component/src/assessment/pages/AssessmentPage/components/AssessmentCompletion/components/GradeSuggestion.tsx
@@ -16,7 +16,7 @@ import { useStudentAssessmentStore } from '../../../../../zustand/useStudentAsse
 import { mapNumberToScoreLevel } from '../../../../../interfaces/scoreLevel'
 
 import { getWeightedScoreLevel } from '../../../../utils/getWeightedScoreLevel'
-import { StudentScoreBadge } from '../../../../components/StudentScoreBadge'
+import { StudentScoreBadge } from '../../../../components/badges'
 
 interface GradeSuggestionProps {
   onGradeSuggestionChange: (value: string) => void
