+++ clients/assessment_component/src/assessment/pages/components/badges/AssessmentStatusBadge.tsx
