+++ clients/assessment_component/src/assessment/pages/components/badges/index.ts
@@ -1,0 +1,6 @@
+export { AssessmentStatusBadge } from './AssessmentStatusBadge'
+export { DeadlineBadge } from './DeadlineBadge'
+export { GradeSuggestionBadge } from './GradeSuggestionBadge'
+export { PeerEvaluationCompletionBadge } from './PeerEvaluationCompletionBadge'
+export { StudentScoreBadge } from './StudentScoreBadge'
+export { TeamBadge } from './TeamBadge'
